version con ERROR RARO resuelto y con todos los servicios y repositorios agregados

+++ platzi-market/src/main/java/com/platzimarket/persistence/entity/Cliente.java
@@ -14,7 +14,7 @@ public class Cliente {
 
     private String apellido;
 
-    private Integer celular;
+    private Long celular;
 
     private String direccion;
 
@@ -48,11 +48,11 @@ public class Cliente {
         this.apellido = apellido;
     }
 
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
